requirements.txt
discord2=2.0.1